Abstract some common methods to AbstractMetadataStore
